fixup rename workflow mappers directory
