+++ pkg/tcl/mapperstcl/testworkflows/mappers_test.go
@@ -6,7 +6,7 @@
 //
 //	https://github.com/kubeshop/testkube/blob/main/licenses/TCL.txt
 
-package mappers
+package testworkflows
 
 import (
 	"testing"
